feat(opencode): refactor opencode-sessions into general OpenCode extension
