--- extensions/opencode-sessions/CHANGELOG.md
@@ -1,10 +1,0 @@
-# OpenCode Sessions Changelog
-
-## [Initial Version] - 2026-02-20
-
-- Browse sessions with time-grouped sections and project filtering
-- View full conversation transcripts with tool usage details
-- Generate AI-powered session summaries
-- Copy transcripts, session IDs, slugs, and resume commands
-- Open project directories and share links
-- Delete individual sessions or all sessions for a project
