--- extensions/opencode-sessions/package.json
@@ -1,53 +1,0 @@
-{
-  "$schema": "https://www.raycast.com/schemas/extension.json",
-  "name": "opencode-sessions",
-  "title": "OpenCode Sessions",
-  "description": "Browse, search, and manage your OpenCode sessions",
-  "icon": "extension-icon.png",
-  "author": "mike182uk",
-  "platforms": [
-    "macOS"
-  ],
-  "categories": [
-    "Developer Tools"
-  ],
-  "license": "MIT",
-  "commands": [
-    {
-      "name": "index",
-      "title": "List Sessions",
-      "description": "Browse all OpenCode sessions with search and filtering",
-      "mode": "view"
-    }
-  ],
-  "preferences": [
-    {
-      "name": "databasePath",
-      "type": "textfield",
-      "required": false,
-      "title": "Database Path",
-      "description": "Override OpenCode database path (defaults to ~/.local/share/opencode/opencode.db)",
-      "placeholder": "~/.local/share/opencode/opencode.db"
-    }
-  ],
-  "dependencies": {
-    "@raycast/api": "^1.104.5",
-    "@raycast/utils": "^1.17.0"
-  },
-  "devDependencies": {
-    "@raycast/eslint-config": "^2.0.4",
-    "@types/node": "22.13.10",
-    "@types/react": "19.0.10",
-    "eslint": "^9.22.0",
-    "prettier": "^3.5.3",
-    "typescript": "^5.8.2"
-  },
-  "scripts": {
-    "build": "ray build",
-    "dev": "ray develop",
-    "fix-lint": "ray lint --fix",
-    "lint": "ray lint",
-    "prepublishOnly": "echo \"\\n\\nIt seems like you are trying to publish the Raycast extension to npm.\\n\\nIf you did intend to publish it to npm, remove the \\`prepublishOnly\\` script and rerun \\`npm publish\\` again.\\nIf you wanted to publish it to the Raycast Store instead, use \\`npm run publish\\` instead.\\n\\n\" && exit 1",
-    "publish": "npx @raycast/api@latest publish"
-  }
-}
