+++ extensions/opencode/.gitignore
@@ -8,8 +8,7 @@ raycast-env.d.ts
 .raycast-swift-build
 .swiftpm
 compiled_raycast_swift
+compiled_raycast_rust
 
 # misc
 .DS_Store
-.idea
-.vscode
