+++ extensions/opencode/CHANGELOG.md
@@ -1,8 +1,17 @@
 # OpenCode Changelog
 
-## [Initial Version] - {PR_MERGE_DATE}
+## [New Feature] - {PR_MERGE_DATE}
 
-### Added
-- Added `Sessions` command to view and manage OpenCode sessions.
-- Added `Commands` command to view and create custom OpenCode commands.
-- Added `Usage Cheatsheet` command for quick reference of TUI, CLI, and Web commands.
+- Renamed extension to OpenCode.
+- Added `OpenCode Commands` to view, filter (All, System, Custom), and create custom commands via Form UI.
+- Added `OpenCode Usage Cheatsheet` for quick reference of TUI, CLI, and Web commands.
+- Enhanced existing session listing to `OpenCode Sessions`.
+
+## [Initial Version] - 2026-02-20
+
+- Browse sessions with time-grouped sections and project filtering
+- View full conversation transcripts with tool usage details
+- Generate AI-powered session summaries
+- Copy transcripts, session IDs, slugs, and resume commands
+- Open project directories and share links
+- Delete individual sessions or all sessions for a project
