+++ extensions/opencode/README.md
