+++ extensions/opencode/assets/extension-icon@dark.png
