+++ extensions/opencode/docs/opencode.md
