+++ extensions/opencode/eslint.config.js
