+++ extensions/opencode/package.json
@@ -2,9 +2,15 @@
   "$schema": "https://www.raycast.com/schemas/extension.json",
   "name": "opencode",
   "title": "OpenCode",
-  "description": "Quickly open your projects in your favorite code editor",
+  "description": "Manage your OpenCode sessions, commands, and more",
   "icon": "extension-icon.png",
-  "author": "SandeepBaskaran",
+  "author": "mike182uk",
+  "contributors": [
+    "SandeepBaskaran"
+  ],
+  "platforms": [
+    "macOS"
+  ],
   "categories": [
     "Developer Tools"
   ],
@@ -12,43 +18,53 @@
   "commands": [
     {
       "name": "sessions",
-      "title": "Sessions",
-      "description": "View and manage your OpenCode sessions",
+      "title": "OpenCode Sessions",
+      "description": "Browse and manage your OpenCode sessions",
       "mode": "view"
     },
     {
       "name": "commands",
-      "title": "Commands",
+      "title": "OpenCode Commands",
       "description": "View and create custom OpenCode commands",
       "mode": "view"
     },
     {
       "name": "cheatsheet",
-      "title": "Usage Cheatsheet",
-      "description": "Cheatsheet for OpenCode TUI, CLI, and Web commands",
+      "title": "OpenCode Usage Cheatsheet",
+      "description": "Quick reference for OpenCode TUI, CLI, and Web commands",
       "mode": "view"
     }
   ],
+  "preferences": [
+    {
+      "name": "databasePath",
+      "type": "textfield",
+      "required": false,
+      "title": "Database Path",
+      "description": "Override OpenCode database path (defaults to ~/.local/share/opencode/opencode.db)",
+      "placeholder": "~/.local/share/opencode/opencode.db"
+    }
+  ],
   "dependencies": {
-    "@raycast/api": "^1.84.3",
+    "@raycast/api": "^1.104.5",
     "@raycast/utils": "^1.17.0",
     "uuid": "^13.0.0"
   },
   "devDependencies": {
-    "@raycast/eslint-config": "^1.0.11",
-    "@types/node": "20.8.10",
-    "@types/node-fetch": "^2.6.13",
-    "@types/react": "18.3.3",
+    "@raycast/eslint-config": "^2.0.4",
+    "@types/node": "22.13.10",
+    "@types/react": "19.0.10",
     "@types/uuid": "^10.0.0",
-    "eslint": "^8.57.0",
-    "prettier": "^3.3.3",
-    "typescript": "^5.4.5"
+    "eslint": "^9.22.0",
+    "prettier": "^3.5.3",
+    "typescript": "^5.8.2"
   },
   "scripts": {
     "build": "ray build",
     "dev": "ray develop",
     "fix-lint": "ray lint --fix",
     "lint": "ray lint",
+    "prepublishOnly": "echo \"\\n\\nIt seems like you are trying to publish the Raycast extension to npm.\\n\\nIf you did intend to publish it to npm, remove the \\`prepublishOnly\\` script and rerun \\`npm publish\\` again.\\nIf you wanted to publish it to the Raycast Store instead, use \\`npm run publish\\` instead.\\n\\n\" && exit 1",
     "publish": "npx @raycast/api@latest publish"
   }
 }
