+++ extensions/opencode/src/components/ProjectDropdown.tsx
