+++ extensions/opencode/src/components/SessionActions.tsx
