+++ extensions/opencode/src/components/SessionDetail.tsx
