+++ extensions/opencode/src/components/SessionListItem.tsx
