+++ extensions/opencode/src/components/SessionSummary.tsx
