+++ extensions/opencode/src/hooks/useSessions.ts
