+++ extensions/opencode/src/lib/storage.ts
