+++ extensions/opencode/src/utils.ts
